Projet_temperature_terrestre/streamlit/requirements.txt
seaborn==0.12.2
scikit-learn==1.2.2
markdown-it-py==2.2.0
pandas==2.0.1
pdf2img==0.1.2
Pillow==9.5.0
plotly-express==0.4.1
scipy==1.10.1
streamlit==1.22.0
streamlit-option-menu==0.3.4
geopandas==0.13.2
wordcloud==1.9.2
matplotlib==3.8.0